Refactor RichTextTag dir
